ref/ui patch

+++ app/src/main/java/is/hi/hbv601g/brent/activities/model/CartActivity.java
@@ -112,7 +112,7 @@ public class CartActivity extends CurrentActivity {
         List<Tour> tours = mCart.getTours();
 
         Map<String, Integer> quantity = new HashMap<>();
-        Map<String, Integer> priceList = new HashMap<>();
+        Map<String, Long> priceList = new HashMap<>();
         Map<String, String> productName = new HashMap<>();
 
         ArrayList<Triplet> listOfTriplets = new ArrayList<>();
@@ -146,7 +146,7 @@ public class CartActivity extends CurrentActivity {
             }
         }
 
-        for (Map.Entry<String, Integer> entry : priceList.entrySet()) {
+        for (Map.Entry<String, Long> entry : priceList.entrySet()) {
             String bikeID = entry.getKey();
             Log.d("Cart", bikeID + " / " + entry.getValue());
             listOfTriplets.add(new Triplet(productName.get(bikeID), quantity.get(bikeID).intValue(), priceList.get(bikeID).intValue()));
